Initial CLI implementation
