+++ video-search/utils.py
@@ -1,3 +1,4 @@
+import logging
 import os
 from functools import lru_cache
 
@@ -24,3 +25,23 @@ def count_records(folder: str, tp: str) -> int:
         ret += 1
 
     return ret
+
+
+def create_logger(name: str, log_file: str) -> logging.Logger:
+    log = logging.getLogger(name)
+    log.setLevel(logging.INFO)
+
+    handler = logging.StreamHandler()
+    handler.setLevel(logging.INFO)
+    formatter = logging.Formatter("%(levelname)s - %(message)s")
+    handler.setFormatter(formatter)
+    log.addHandler(handler)
+
+    # create error file handler and set level to error
+    handler = logging.FileHandler(log_file, "w")
+    handler.setLevel(logging.INFO)
+    formatter = logging.Formatter("%(levelname)s - %(message)s")
+    handler.setFormatter(formatter)
+    log.addHandler(handler)
+
+    return log
